@@ -6,17 +6,18 @@
     "react",
     "starter"
   ],
+  "type": "module",
   "main": "src/index.js",
   "dependencies": {
-    "antd": "4.18.6",
-    "axios": "^1.3.2",
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "antd": "4.24.15",
+    "axios": "^1.6.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
     "@babel/runtime": "7.13.8",
-    "@vitejs/plugin-react": "^2.0.1",
-    "vite": "^3.0.7"
+    "@vitejs/plugin-react": "^4.2.1",
+    "vite": "^5.0.7"
   },
   "scripts": {
     "predeploy": "yarn build",
