@@ -1,12 +1,7 @@
 import { StrictMode } from "react";
-import ReactDOM from "react-dom";
-
 import App from "./App";
+import { createRoot } from "react-dom/client";
 
-const rootElement = document.getElementById("root");
-ReactDOM.render(
-  <StrictMode>
-    <App />
-  </StrictMode>,
-  rootElement
-);
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(<App />);
